project-repo/requirements.txt
aiobotocore==2.21.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
bitarray==3.2.0
boto3==1.37.22
botocore==1.37.1
canonicaljson==2.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.1.0
click==8.1.8
coloredlogs==14.0
cytoolz==1.0.1
dagster==1.10.6
dagster-graphql==1.10.6
dagster-pipes==1.10.6
dagster-shared==0.26.6
dagster-webserver==1.10.6
docstring_parser==0.16
eth-account==0.13.6
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
exceptiongroup==1.2.2
farcaster==0.7.12
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
gql==3.5.2
graphene==3.4.3
graphql-core==3.2.4
graphql-relay==3.2.0
grpcio==1.71.0
grpcio-health-checking==1.71.0
h11==0.14.0
hexbytes==1.3.0
httptools==0.6.4
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.2.0
neo4j==5.28.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
propcache==0.3.1
protobuf==5.29.4
pycryptodome==3.22.0
pydantic==2.11.0
pydantic_core==2.33.0
Pygments==2.19.1
pyhumps==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rlp==4.1.0
s3fs==2025.3.0
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
structlog==25.2.0
tabulate==0.9.0
tomli==2.2.1
toolz==1.0.0
toposort==1.10
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
universal_pathlib==0.2.6
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchdog==5.0.3
watchfiles==1.0.4
websockets==15.0.1
wrapt==1.17.2
yarl==1.18.3